Visualize the shortest path with Leaflet

The function path_leaflet visualizes the shortest path (in minutes or meters) that has been calculated with the function shortest_path_igraph() (where path = TRUE). Internet connection is required to load the background tiles.

Usage

path_leaflet(path, graph_object = graph)

Arguments

path

Object (list) that has been created with the function shortest_path_igraph(), where path = TRUE.

graph_object

The road network structured as a tidy graph (tbl_graph object). This can be done with the function vegnett_to_R().

Value

Interactive Leaflet map that shows the shortest path (in minutes or meters) between a chosen from and to node in the road network.

Examples

vegnett_list <- vegnett_to_R(vegnett = vegnett_sampledata,
                             year = 2021,
                             fromnodeID = "FROMNODEID",
                             tonodeID = "TONODEID",
                             FT_minutes = "FT_MINUTES",
                             TF_minutes = "TF_MINUTES",
                             meters = "SHAPE_LENGTH")

graph <- vegnett_list[[1]]
nodes <- vegnett_list[[2]]
edges <- vegnett_list[[3]]
graph_cppRouting_minutes <- vegnett_list[[4]]
graph_cppRouting_meters <- vegnett_list[[5]]

path <- shortest_path_igraph(from_node_ID = 25,
                             to_node_ID = 33,
                             unit = "minutes",
                             path = TRUE)

path_leaflet(path)