Convert the Norwegian road network (NVDB Ruteplan nettverksdatasett) into network graphs in R

The function vegnett_to_R can be used to convert the Norwegian road network, downloaded from Geonorge, to formats that allows for network analysis in R (tbl_graph and cppRouting).

Usage

vegnett_to_R(vegnett, crs_out = 25833)

Arguments

vegnett

The Norwegian road network, downloaded from Geonorge, as an sf object.

crs_out

Numeric vector with the chosen coordinate reference system (CRS).

Value

List containing the following elements:

[1] graph: the road network structured as a tidy graph (tbl_graph object)

[2] nodes: the road network's nodes (sf object)

[3] edges: road network's edges/node links (data.frame)

[4] graph_cppRouting_FT_MINUTES: the road network structured as a cppRouting graph with the cost of travel in minutes (cppRouting object)

[5] graph_cppRouting_LENGTH: the road network structured as a cppRouting graph with the cost of travel in meters (cppRouting object)

Examples

vegnett_sampledata
#> Simple feature collection with 1000 features and 45 fields
#> Geometry type: MULTILINESTRING
#> Dimension:     XYZ
#> Bounding box:  xmin: 248861 ymin: 6639120 xmax: 273377.4 ymax: 6672977
#> z_range:       zmin: -999999 zmax: 448.88
#> Projected CRS: ETRS89 / UTM zone 33N
#> # A tibble: 1,000 x 46
#>    ROUTEID FROM_M  TO_M FromNod~1 ToNod~2 GATEN~3 VEGTYPE ONEWAY FT_MI~4 TF_MI~5
#>    <chr>    <dbl> <dbl> <chr>     <chr>   <chr>   <chr>   <chr>    <dbl>   <dbl>
#>  1 1209471  0     0.807 1209475   2673001 Rv150   R       FT     0.0508  0.0508 
#>  2 1209471  0.807 1     2673001   1209476 Rv150   R       FT     0.0125  0.0125 
#>  3 1209472  0     0.218 1209477   2673002 Rv150   R       FT     0.0138  0.0138 
#>  4 1209472  0.218 1     2673002   1209478 Rv150   R       FT     0.0517  0.0517 
#>  5 1209473  0     0.129 627807    3066146 Rv150   R       FT     0.00755 0.00755
#>  6 1209473  0.129 1     3066146   1209475 Rv150   R       FT     0.0399  0.0399 
#>  7 1209474  0     0.828 1209478   3278307 Rv150   R       FT     0.0425  0.0425 
#>  8 1209474  0.828 1     3278307   627806  Rv150   R       FT     0.00883 0.00883
#>  9 1209419  0     1     1209425   1209426 Rv163   R       FT     0.0993  0.0993 
#> 10 1209420  0     1     1209426   1209427 Rv163   R       FT     0.174   0.174  
#> # ... with 990 more rows, 36 more variables: METER_ORG <dbl>, VEGNUMMER <int>,
#> #   FYLKE_ID <chr>, HIERARCHY <int>, HIERARCHY_FVK <int>, FuncRoadClass <int>,
#> #   VEGID <chr>, SPECIALVEG <chr>, URL <chr>, ATTRIBUTES <chr>,
#> #   FT_BOM_LIT <dbl>, FT_BOM_STO <dbl>, TF_BOM_LIT <dbl>, TF_BOM_STO <dbl>,
#> #   SPERRING <int>, DT_TURIST_FT <dbl>, DT_TURIST_TF <dbl>, DT_TRUCK_FT <dbl>,
#> #   DT_TRUCK_TF <dbl>, TOTALVIKT <dbl>, SK_HO_FT <dbl>, SK_HO_TF <dbl>,
#> #   VOGNTOGLENGDE <dbl>, RoadClass <int>, Medium <chr>, FT_Fart <int>, ...
#> # i Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
vegnett_list <- vegnett_to_R(vegnett = vegnett_sampledata)
#> Error in dplyr::select(., from, to, SHAPE_LENGTH): Can't subset columns that don't exist.
#> x Column `SHAPE_LENGTH` doesn't exist.

graph <- vegnett_list[[1]]
#> Error in eval(expr, envir, enclos): object 'vegnett_list' not found
nodes <- vegnett_list[[2]]
#> Error in eval(expr, envir, enclos): object 'vegnett_list' not found
edges <- vegnett_list[[3]]
#> Error in eval(expr, envir, enclos): object 'vegnett_list' not found
graph_cppRouting_FT_MINUTES <- vegnett_list[[4]]
#> Error in eval(expr, envir, enclos): object 'vegnett_list' not found
graph_cppRouting_LENGTH <- vegnett_list[[5]]
#> Error in eval(expr, envir, enclos): object 'vegnett_list' not found

graph
#> Error in eval(expr, envir, enclos): object 'graph' not found
nodes
#> Error in eval(expr, envir, enclos): object 'nodes' not found
head(edges)
#> Error in head(edges): object 'edges' not found
head(graph_cppRouting_FT_MINUTES$data)
#> Error in head(graph_cppRouting_FT_MINUTES$data): object 'graph_cppRouting_FT_MINUTES' not found
head(graph_cppRouting_FT_MINUTES$coords)
#> Error in head(graph_cppRouting_FT_MINUTES$coords): object 'graph_cppRouting_FT_MINUTES' not found
head(graph_cppRouting_FT_MINUTES$dict)
#> Error in head(graph_cppRouting_FT_MINUTES$dict): object 'graph_cppRouting_FT_MINUTES' not found
graph_cppRouting_FT_MINUTES$nbnode
#> Error in eval(expr, envir, enclos): object 'graph_cppRouting_FT_MINUTES' not found

head(graph_cppRouting_LENGTH$data)
#> Error in head(graph_cppRouting_LENGTH$data): object 'graph_cppRouting_LENGTH' not found
head(graph_cppRouting_LENGTH$coords)
#> Error in head(graph_cppRouting_LENGTH$coords): object 'graph_cppRouting_LENGTH' not found
head(graph_cppRouting_LENGTH$dict)
#> Error in head(graph_cppRouting_LENGTH$dict): object 'graph_cppRouting_LENGTH' not found
graph_cppRouting_LENGTH$nbnode
#> Error in eval(expr, envir, enclos): object 'graph_cppRouting_LENGTH' not found